# onvif/api/internal/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "接口文档",
        "contact": {},
        "version": "1.0.0"
    },
    "paths": {
...
        "/media_servers": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "流媒体服务"
                ],
                "summary": "列表",
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "条数，小于 1 不匹配",
                        "name": "count",
                        "in": "query"
                    },
                    {
                        "enum": [
                            0,
                            1
                        ],
                        "type": "integer",
                        "description": "是否在线，精确",
                        "name": "describe",
                        "in": "query"
                    },
                    {
                        "enum": [
                            0,
                            1
                        ],
                        "type": "integer",
                        "description": "是否禁用，精确",
                        "name": "enable",
                        "in": "query"
                    },
                    {
                        "maxLength": 32,
                        "type": "string",
                        "description": "id，模糊",
                        "name": "id",
                        "in": "query"
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "所属的分组，精确",
                        "name": "mediaServerGroupID",
                        "in": "query"
                    },
                    {
                        "maxLength": 32,
                        "type": "string",
                        "description": "名称，模糊",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "minimum": 0,
                        "type": "integer",
                        "description": "偏移，小于 0 不匹配",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "排序，\"column [desc]\"",
                        "name": "order",
                        "in": "query"
                    },
                    {
                        "maxLength": 64,
                        "type": "string",
                        "description": "访问密钥，模糊",
                        "name": "secret",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/util.GORMList-db_MediaServer"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal.Error"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "流媒体服务"
                ],
                "summary": "添加",
                "parameters": [
                    {
                        "description": "数据",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/mediaservers.postReq"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/internal.IDResult-int64"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal.Error"
                        }
                    }
                }
            },
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "流媒体服务"
                ],
                "summary": "批量删除",
                "parameters": [
                    {
                        "description": "id 数组",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal.BatchDelete-string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal.Error"
                        }
                    }
                }
            }
        },
        "/media_servers/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "流媒体服务"
                ],
                "summary": "详情",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/db.MediaServer"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal.Error"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "流媒体服务"
                ],
                "summary": "删除",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal.Error"
                        }
                    }
                }
            },
            "patch": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "流媒体服务"
                ],
                "summary": "修改",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "数据",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/mediaservers.patchReq"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/internal.RowResult"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal.Error"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "db.MediaServer": {
            "type": "object",
            "properties": {
                "apiBaseURL": {
                    "description": "API 地址 (http|https)://ip:port",
                    "type": "string"
                },
                "apiCallTimeout": {
                    "description": "请求服务接口超时时间，单位，毫秒，默认 5000",
                    "type": "integer"
                },
                "enable": {
                    "description": "是否禁用，0/1",
                    "type": "integer"
                },
                "id": {
                    "description": "id",
                    "type": "string"
                },
                "keepaliveTime": {
                    "description": "心跳时间戳",
                    "type": "integer"
                },
                "mediaServerGroupID": {
                    "description": "所属的分组",
                    "type": "integer"
                },
                "name": {
                    "description": "名称，方便记忆",
                    "type": "string"
                },
                "online": {
                    "description": "是否在线，0/1",
                    "type": "integer"
                },
                "privateIP": {
                    "description": "内网访问的 ip ，生成播放地址时使用",
                    "type": "string"
                },
                "publicIP": {
                    "description": "外网访问的 ip ，生成播放地址时使用",
                    "type": "string"
                },
                "secret": {
                    "description": "访问密钥",
                    "type": "string"
                }
            }
        },
...
        "internal.BatchDelete-string": {
            "type": "object",
            "required": [
                "id"
            ],
            "properties": {
                "id": {
                    "description": "数据库 ID 数组",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "internal.Error": {
            "type": "object",
            "properties": {
                "detail": {
                    "description": "详细信息",
                    "type": "string"
                },
                "phrase": {
                    "description": "简短信息",
                    "type": "string"
                }
            }
        },
        "internal.IDResult-int64": {
            "type": "object",
            "properties": {
                "id": {
                    "description": "数据库 ID",
                    "type": "integer"
                }
            }
        },
        "internal.RowResult": {
            "type": "object",
            "properties": {
                "row": {
                    "description": "行数",
                    "type": "integer"
                }
            }
        },
...
        "mediaservers.patchReq": {
            "type": "object",
            "properties": {
                "apiBaseURL": {
                    "description": "API 地址 (http|https)://ip:port",
                    "type": "string"
                },
                "apiCallTimeout": {
                    "description": "请求服务接口超时时间，单位，毫秒，默认 5000",
                    "type": "integer",
                    "minimum": 1000
                },
                "describe": {
                    "description": "描述",
                    "type": "string",
                    "maxLength": 128
                },
                "enable": {
                    "description": "是否禁用，0/1 ，默认 1",
                    "type": "integer",
                    "enum": [
                        0,
                        1
                    ]
                },
                "mediaServerGroupID": {
                    "description": "所属的分组 ，默认 1",
                    "type": "integer",
                    "minimum": 1
                },
                "name": {
                    "description": "名称",
                    "type": "string",
                    "maxLength": 32
                },
                "privateIP": {
                    "description": "内网访问的 ip ，生成播放地址时使用，默认使用 apiBaseURL 中的 host",
                    "type": "string"
                },
                "publicIP": {
                    "description": "外网访问的 ip ，生成播放地址时使用，默认使用 apiBaseURL 中的 host",
                    "type": "string"
                },
                "secret": {
                    "description": "访问密钥",
                    "type": "string",
                    "maxLength": 64
                }
            }
        },
        "mediaservers.postReq": {
            "type": "object",
            "required": [
                "apiBaseURL",
                "id",
                "name",
                "secret"
            ],
            "properties": {
                "apiBaseURL": {
                    "description": "API 地址 (http|https)://ip:port",
                    "type": "string"
                },
                "apiCallTimeout": {
                    "description": "请求服务接口超时时间，单位，毫秒，默认 5000",
                    "type": "integer",
                    "minimum": 1000
                },
                "describe": {
                    "description": "描述",
                    "type": "string",
                    "maxLength": 128
                },
                "enable": {
                    "description": "是否禁用，0/1 ，默认 1",
                    "type": "integer",
                    "enum": [
                        0,
                        1
                    ]
                },
                "id": {
                    "description": "id",
                    "type": "string",
                    "maxLength": 32
                },
                "mediaServerGroupID": {
                    "description": "所属的分组 ，默认 1",
                    "type": "integer",
                    "minimum": 1
                },
                "name": {
                    "description": "名称",
                    "type": "string",
                    "maxLength": 32
                },
                "privateIP": {
                    "description": "内网访问的 ip ，生成播放地址时使用，默认使用 apiBaseURL 中的 host",
                    "type": "string"
                },
                "publicIP": {
                    "description": "外网访问的 ip ，生成播放地址时使用，默认使用 apiBaseURL 中的 host",
                    "type": "string"
                },
                "secret": {
                    "description": "访问密钥",
                    "type": "string",
                    "maxLength": 64
                }
            }
        },
        "util.GORMList-db_MediaServer": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "列表",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/db.MediaServer"
                    }
                },
                "total": {
                    "description": "总数",
                    "type": "integer"
                }
            }
        },
...
    }
}